Initialized project with npm init.
Set up testing environment. The goal of this project is to not only create my own personal website but also to focus on test driven development this time around. R2FC did not utilize test driven development which is not ideal when coding.

Test suite confirmed to be working in the server section.

Next is server side dependencies installation.