Home Page.
Home page basic (tentative) file structure created.

Home page will render two separate children called about-me and projects-display.

About-me will display the description that I decide to store in my database which is updatable via a log-in input.

Projects-display will display a section with cards showing all the projects that I worked on. This projects-display section will have a separate child component that takes in a prop for the current project and displays the information in a "card" format.

Will create the projects-display child section next.