Home-Page-1 | Navigation Bar
Navigation Bar Implemented into home page.

Need to add hrefs to each of the listed navigation items but I will do so once the components are worked on.

Created an additional folder to hold my current navigation bar items. This allows for me to use less redundant code by using a map through the items instead of manually creating each one, assigning the classes to each one, and so on so forth.

Next up is our "About Me" Section.