home-dev-2 | About Me
Basic home page about me created.

Utilized object sample to pull description into the About Me. This makes it so the in the future when I want to update my about me, it's simple to do from the database rather than hard coding an example. I might go as far as to make this a downloadable template/login feature so people can host their own display website.

Size and component css translated to be centered in the parent compononent.

Next up is to create our project card sample objects. Just one to start is fine and then we can create more to check how they're going to be rendered out.