projects-display-dev-1 | className
className naming conventions have been updated.

Since this is a reusbale component with different styling required based on where the component is being rendered, I updated the naming convention to utilize a prop instead of a hard coded className. This allows me to avoid having to do any additional work other than simply passing down the component type as a prop so our css styling applies as needed.

For example, I pass down the prop compType={"home"} to the project display component that exists in the home page. This applies all the css styling I had done previously for the home page.

For the main component project display, I pass down the "main" compType prop and it applies a different set of CSS styling as needed which will be worked on next.