css-dev-1 | home page css
Home Page CSS updated to be at a maximum width of 720 pixels. The width matches up with bootstrap's medium container size but the component won't be responsive past 720 pixels which I found to be a sweet spot. Anything under 720 px is still responsive.

Margins and padding updated to have a very clean look now.