css-dev-1 | project-display
Projects display updated to now match the sizing of the home page at a max-width of 720px. I had made changes to the project display component and how the cards are shown mainly different padding and spacing for the component.

I also updated the reusable component to remove the differentiation between the home page and the main component display page as the look on the home page is nice and I want to reuse the same css for the separate main projects display page.