css-dev-1 | about me
Resume viewer now working. Utilizes an onClick function to open my resume in a new window. Would be based on the user's browser and settings. In Google Chrome it would either download the pdf or it would open it up in a new window.

Basic css styling complete at this point, moving on to enter dark mode only.