css-dev-1 | css complete
General CSS styling and colors complete for Home, About Me, Projects Display, and Project Details.

Router is now working with bundle.js being update to /bundle.js and removing the href. This is important so that we can have our markdown file Table of Contents jump to the sections as necessary.