css-dev-2 | background gradient & font
Designer girlfriend advised me of some better fonts that I could use. Changed the font from system default to Montserrat.

Background gradient added to add some spice to the design.

Next is to refactor the project cards to use an image instead of just the title and the description.

Title will probably remain but a refactor of the description will have to happen. May add icons next to demo details and source.