css-dev-2 | navigation links
Navigation bar links now highlight the same color as the headings.

At first, I was going to use react states to keep track of whether the link was active or not. I ran into the issue though of needing to refactor the map to add unique states for each of the different links. However, I found that instead of having states and writing more code, I could use the active attribute on the NavItem element and feed it a boolean. I used a simple window.location to determine the file path it's currently on and the href provided in the currentLink object to provide a true or false.

This allows for a uniqe active className to be added to the navigation link when it is selected as the current window will match the nav link's href.

A simple css addition for color then reflects the desired effect of a highlighted link.