css-dev-2 | in progress
An In Progress Component has been added in place of a "error boundary" component in react router. The reason for this is because there isn't necessarily and error that occurs when clicking journal just that the component to be rendered hasn't been developed yet. I would have to add the in progress component to each of the current routes which is just more code instead of a temporary component that's linked to the route.

This actually makes it easier in the future as I would just need to swap out the element being rendered one time.