css-dev-2 | footer
footer created and completed.

Home page about me section updated to include my socials instead of putting in the footer for more exposure.

Also three rows in the footer made it thicker than I'd like so left it with just two rows.

Basic css complete at this point. The only refactoring that needs to be done is to create my server file to replace the dev server with the hot reload function and to create a component function to grab out readme file on component render.