project-display-dev-2
Changing how the projects are being displayed. Currently homepage project display revamped to include a website preview image.

On hover, the preview image disappears and is converted to the prior project card.

Most of the work was done through css. Will check the main project display component page to see if function still works. No further commits would be that it indeed does.