project-display-dev-3 | Major Refactor
To give more highlight/emphasis onto the projects display section, I decided to increase the size of each project card to have it's own section to be beautiful (based on advice from engineers/designers).

Major factor needs to be made to the overlay. ClassNames have been update to reflect the new changes such as "overlay."

Working on different viewpoint max heights/widths for the overlay to be properly responsive on different devices. Not possible to fully make it responsive due to component constraints but should fit all use cases for the time being.