project-display-dev-3 | overlay CSS update
Project display component css updated.

Prior css was working fine, but I decided to change the cards from small to large to emphasize my HQ projects more.

Custom viewport width css has been added for the overlay to match up with with responsive components that it's covering. Sadly I wasn't able to figure out a method that allowed an absolute positioned element to naturally follow it's prior parent (not sure if it's possible) so I created my own custom responsive css for that component on all devices available on devtools.