client-server-dev-1 | Database connection to Client
Client is now able to pull up the basic project information from our database to render the projects we have stores.

Currently working on setting up the project-details page to pull the readme links from the database.

Working on setting up a route that servers both the page itself while pulling project readme data based on the second route parameter of the ID.