client-server-dev-1 | Project Details Connection
After some investigation, I found that I'm unable to send data as well as serve the absolute file path to the client side for routing there.

Thus, we need to find another option. I found that I can attach query parameters to the request for readme data to grab the associated readme with our project id