client-server-dev-1 | Project Details
Project Details is now able to pull from our database.

Running a useEffect on componnent load, the useEffect takes the return value from the custom hook "useParams" associated with react-router-dom to parse our URL and obtain our unique ID.

We then send a get request to our server's API which takes the unique id and makes a request to our database to return the github raw readme file url.

This removes the necessity for the readme's to be held in a separate file and works without our samples now.