nav-refactor-2 | Refactor CSS
Similar to the footer, the navigation bar looks good, but when we get to smaller device views the nav bar would shrink and form a new line.

This is fine but I refactored it so at those smaller widths, the overall size of the font on all components would be reduced. This shrinks the space needed for the navigation bar items and makes it look better responsively.

Aside from that, I updated the about-me section of the navigation bar and the url path to simply "about." About me was a bit too long to fit everything in the navigation bar and it looks just as clean now.

Next up is to create our journal entry section!