journal-dev-1 | Editor Unlock
Although this is going to be an undisclosed path, it's still possible that people can just read through my source code and figure out the path that's going to be represent my personal journal entry editor.

To add a second measure of security, I implemented a password section.

This requires a password entry that only I will "know" and require users to enter the associated password to unlock the editor.