journal-dev-1 | Update Path
Updated path for journal-editor may update this path again in the near future as an environment variable so people won't be able to see the actual path checked into github.

This is probably a better idea to make sure that my endpoints are secure.

Will update again in the future when that is implemented.