journal-dev-1 | Functionality / CSS
Journal functionality complete.

Journal now displays all entries and has a highlight feature when hovering over the specific entry.

When an entry is clicked, it redirects to the entry page which displays the markdown text the same as the project details page would.

I thought about creating a reusable component for this since they both use markdown text and I may go ahead and do so in the future but due to the different routes and database methods, I have no plans of doing so at the moment.

Only update that needs to be made is to create a heading for the entry main page.