deployment-refactor-4 | CSS Update
The project display component nesting differed between the home page and the actual project display main component.

Due to the container nesting being different, I had to remove the padding from the home-parent component so it would match the project display main component. Then add additional css to the root div so that we wouldn't have crazy margins on mobile.